[+BUGFIX] smaller deprecated clean-ups
As a backport Extbase is influenced by all the changes in FLOW. It's
important to keep both in sync and thus deprecate code which is not
in FLOW3 anymore or which has changed.

All deprecated annotations should now match the CGL and all removals
are terminated in the future.

* clean up deprecation annotation

Change-Id: I442e1e2d2d0968c4116e2cbf7c77bb5d76f9e3ee
Fixes: #27148
Relates: #13745