[+BUGFIX] Avoiding cHash for nonCacheable actions prevent action from being called
With #8565 cHash is only appended for calls to cachable actions
This has the side effect, that calling non-cacheable actions
from a cached action on the same page fails to execute the
target action.
This fix reverts the changes from #8565 as we couldn't find
a better solution for now. See #13970 for details.

Change-Id: I68c19feb3027bcf578dc4836625a5f98494252d4
Related: #13970